Fix: Searching in OpenSubtitles.org and Napisy24 was not working for files larger then 2GiB

+++ libqnapi/src/engines/napisy24downloadengine.cpp
@@ -60,7 +60,7 @@ QString Napisy24DownloadEngine::checksum(QString filename) {
        i < 65536 / sizeof(tmp) && file.read((char *)&tmp, sizeof(tmp));
        i++, hash += tmp)
     ;
-  file.seek(qMax(0, (int)((qint64)fileSize - 65536)));
+  file.seek(qMax((qint64)0, (qint64)fileSize - 65536));
   for (tmp = 0, i = 0;
        i < 65536 / sizeof(tmp) && file.read((char *)&tmp, sizeof(tmp));
        i++, hash += tmp)
